b-rfc: move nthAlphabet to utils
